requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-colorfield==0.8.0
Pillow==9.3.0
sqlparse==0.4.3
uuid==1.30
views==0.3
